Update unit tests to support custom config

+++ tests/minimal_test.sh
@@ -1,21 +1,29 @@
 set -o errexit 
 set -o pipefail
+if [ -z $1 ];then 
+
+    # Standard mini-test with wiki25
+
+    config=configs/wiki25.sh  
+
+    # Delete previous runs is exist
+    rm -Rf DATA/wiki25/*
+
+    # replace code above with less restrictive deletion
+    # rm -R -f DATA/wiki25/embeddings
+    # rm -R -f DATA/wiki25/features
+    # rm -R -f DATA/wiki25/oracles
+    # rm -R -f DATA/wiki25/models
+
+    # simulate completed corpora extraction and alignment
+    bash tests/create_wiki25_mockup.sh
+
+else
+
+    # custom config mini-test
+    config=$1
+fi
 set -o nounset 
-
-# config=configs/wiki25-neur-al-sampling.sh
-config=configs/wiki25.sh  
-
-# Delete previous runs is exist
-rm -Rf DATA/wiki25/*
-
-# simulate completed corpora extraction and alignment
-bash tests/create_wiki25_mockup.sh
-
-# replace code above with less restrictive deletion
-# rm -R -f DATA/wiki25/embeddings
-# rm -R -f DATA/wiki25/features
-# rm -R -f DATA/wiki25/oracles
-# rm -R -f DATA/wiki25/models
 
 bash run/run_experiment.sh $config
